network: Refactor network.go
Split endpoint implementations into their own file.

Fixes #799
