+++ virtcontainers/endpoint.go
@@ -1,0 +1,87 @@
+// Copyright (c) 2018 Intel Corporation
+//
+// SPDX-License-Identifier: Apache-2.0
+//
+
+package virtcontainers
+
+import (
+	"fmt"
+)
+
+// Endpoint represents a physical or virtual network interface.
+type Endpoint interface {
+	Properties() NetworkInfo
+	Name() string
+	HardwareAddr() string
+	Type() EndpointType
+	PciAddr() string
+	NetworkPair() *NetworkInterfacePair
+
+	SetProperties(NetworkInfo)
+	Attach(hypervisor) error
+	Detach(netNsCreated bool, netNsPath string) error
+	HotAttach(h hypervisor) error
+	HotDetach(h hypervisor, netNsCreated bool, netNsPath string) error
+}
+
+// EndpointType identifies the type of the network endpoint.
+type EndpointType string
+
+const (
+	// PhysicalEndpointType is the physical network interface.
+	PhysicalEndpointType EndpointType = "physical"
+
+	// VirtualEndpointType is the virtual network interface.
+	VirtualEndpointType EndpointType = "virtual"
+
+	// VhostUserEndpointType is the vhostuser network interface.
+	VhostUserEndpointType EndpointType = "vhost-user"
+
+	// BridgedMacvlanEndpointType is macvlan network interface.
+	BridgedMacvlanEndpointType EndpointType = "macvlan"
+
+	// MacvtapEndpointType is macvtap network interface.
+	MacvtapEndpointType EndpointType = "macvtap"
+)
+
+// Set sets an endpoint type based on the input string.
+func (endpointType *EndpointType) Set(value string) error {
+	switch value {
+	case "physical":
+		*endpointType = PhysicalEndpointType
+		return nil
+	case "virtual":
+		*endpointType = VirtualEndpointType
+		return nil
+	case "vhost-user":
+		*endpointType = VhostUserEndpointType
+		return nil
+	case "macvlan":
+		*endpointType = BridgedMacvlanEndpointType
+		return nil
+	case "macvtap":
+		*endpointType = MacvtapEndpointType
+		return nil
+	default:
+		return fmt.Errorf("Unknown endpoint type %s", value)
+	}
+}
+
+// String converts an endpoint type to a string.
+func (endpointType *EndpointType) String() string {
+	switch *endpointType {
+	case PhysicalEndpointType:
+		return string(PhysicalEndpointType)
+	case VirtualEndpointType:
+		return string(VirtualEndpointType)
+	case VhostUserEndpointType:
+		return string(VhostUserEndpointType)
+	case BridgedMacvlanEndpointType:
+		return string(BridgedMacvlanEndpointType)
+	case MacvtapEndpointType:
+		return string(MacvtapEndpointType)
+	default:
+		return ""
+	}
+}
